moving global __init__ functionality to io_utils.py
